Moved `follow` / `unfollow` API actions to dedicated controller
no ref

Moved `follow` / `unfollow` API actions to a dedicated controller to follow
the same pattern as the `block` controller, in preparation for modification of
the `handleFollow` to prevent processing of requests from blocked accounts
